ai-orchestrator/requirements.txt
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
flask>=2.0.0